requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
playwright
email-validator
uploads/
Pillow
